🐛 fix(model-pools): ignore stale stream callbacks
